tests: Move token-contract tests into contract workspace member
