--- tests/src/common/mod.rs
@@ -1,12 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-//
-// Copyright (c) DUSK NETWORK. All rights reserved.
-
-//! Taken from rusk transfer-contract tests.
-//! TODO: These utils could be generally available somewhere
-
-pub mod network;
-pub mod session;
-pub mod utils;
